Modified index.html
I have done three things with this update.

I have updated my status as a student, to where I am now.

I have removed the reference to lahs.club, as I am no longer affiliated with the club.

I have modified the copyright to get people to the theme that this site is based off faster.